\name{RSFTP-package}
\alias{RSFTP-package}
\alias{RSFTP}
\docType{package}
\title{
Tools for Communicating via SSH File Transfer Protocol (SFTP)
}
\description{
This package provides secure-shell (ssh) access to remote servers.  Version 1.0 has just six commands, for uploading, downloading and 
deleting files, and for creating, checking and deleting directories, all via SFTP.  The package wraps the source code for PuTTY.  As
such it could be extended to provide additional ssh features.
}
\details{
\tabular{ll}{
Package: \tab RSFTP\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2014-12-12\cr
License: \tab  GPL-3\cr
}
This library is a wrapper around PuTTY, using its source code according to the license terms specified here: http://www.chiark.greenend.org.uk/~sgtatham/putty/licence.html

    PuTTY is copyright 1997-2014 Simon Tatham.

    Portions copyright Robert de Bath, Joris van Rantwijk, Delian Delchev, Andreas Schultz, Jeroen Massar, Wez Furlong, Nicolas Barry, Justin Bradford, Ben Harris, Malcolm Smith, Ahmad Khalifa, Markus Kuhn, Colin Watson, and CORE SDI S.A.

    Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

    The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL SIMON TATHAM BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. 


}
\author{
Bruce Hoff <bruce.hoff@sagebase.org>

Maintainer: Bruce Hoff <bruce.hoff@sagebase.org>

}

\references{


http://www.chiark.greenend.org.uk/~sgtatham/putty/


}
\keyword{ package }

